pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
authors = [{name = "Your Name", email = "your@email.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0"
]

[project.scripts]
gendiff = "gendiff.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=5.0.0",
    "autopep8>=2.0.0"
]

[tool.setuptools]
packages = ["gendiff"]
